+++ packages/localizations/src/fr-FR.ts
@@ -1769,6 +1769,7 @@ export const frFR: LocalizationResource = {
     form_password_compromised__sign_in: undefined,
     form_password_incorrect: 'Mot de passe incorrect',
     form_password_length_too_short: 'Votre mot de passe est trop court.',
+    form_password_matches_identifier: undefined,
     form_password_not_strong_enough: "Votre mot de passe n'est pas assez fort.",
     form_password_or_identifier_incorrect:
       "Le mot de passe ou l'adresse e-mail est incorrect. Réessayez ou utilisez une autre méthode.",