+++ packages/expo-passkeys/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @clerk/expo-passkeys
 
+## 2.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`7f5c294`](https://github.com/clerk/javascript/commit/7f5c2947e2b3b2ac9677116ff7eede61a1dab649)]:
+  - @clerk/shared@4.29.1
+
 ## 2.0.10
 
 ### Patch Changes